pyproject.toml
[project]
name = "finance-purple-agent"
version = "0.1.0"
description = "Baseline agent for the finance benchmark"
authors = [
    {name = "Anjana Sarkar"},
    {name = "Uthira Mohan"},
    {name = "Thomas Plunkett"},
    {name = "Iraitz Montalbán"},

]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "a2a-sdk[http-server]>=0.3.20",
    "pydantic>=2.0.0",
    "uvicorn>=0.27.0",
    "starlette>=0.35.0",
    "python-dotenv>=1.0.0",
    "fastmcp>=2.0.0",
    "loguru>=0.7.3",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "ruff>=0.14.11",
]
